refactor: Point chat graph at single_shot.pipeline directly (#21)
- chat/graph/state.py imports QueryResult from
  chat.services.single_shot.types (was query_pipeline).
- chat/graph/app.py imports QueryResult + QueryPipelineError from
  the same location.
- chat/graph/nodes/single_shot.py now calls run_single_shot
  directly; no longer depends on query_pipeline.answer_question.

The query_pipeline shim only has the Phase 2 feature-flag direct
path left as a consumer, which will be removed in the next commit.
